Added proper frontend, added calendar aware filtering and proper arrival implementation aswell as new 'trip updates' feed integrated across. docker compose set up correctly. added readme with project detaisl and ascii diagram

+++ backend/src/main/java/dev/aroncalvert/gtfspulse/dto/ArrivalDTO.java
@@ -3,5 +3,5 @@ package dev.aroncalvert.gtfspulse.dto;
 public record ArrivalDTO(
     String routeShortName,
     String headsign,
-    int arrivalTime) {
+    long arrivalTime) {
 }
